Partially resolved git commit message problem
It seems that for anything beyond a very simple change, we are going to
be omitting the -m option for "git commit". By doing so, we get into
VSCode and we have a neat little line on the right that seems to be the
recommended limit of 72 characters.

Pretty cool! Denicia got stuck near the end of Pony Island and is now
laying on my futon :3